+++ lib/alf/renderer.rb
@@ -130,3 +130,4 @@ end # module Alf
 require_relative 'renderer/rash'
 require_relative 'renderer/text'
 require_relative 'renderer/yaml'
+require_relative 'renderer/json'